wrap the app in an errorboundary as a last line catch for Sentry.
