+++ app/src/app/index.jsx
@@ -1,13 +1,11 @@
-/* eslint-disable class-methods-use-this */
 // @flow
 
 import '$shared/assets/stylesheets'
 
-import React, { Component } from 'react'
+import React from 'react'
 import { Route as RouterRoute, Switch, Redirect, type Location } from 'react-router-dom'
 import { ConnectedRouter } from 'react-router-redux'
 import qs from 'query-string'
-import * as Sentry from '@sentry/browser'
 
 // Marketplace
 import ProductPage from '$mp/containers/ProductPage'
@@ -64,8 +62,9 @@ import GoogleAnalyticsTracker from '$mp/components/GoogleAnalyticsTracker'
 import isProduction from '$mp/utils/isProduction'
 import ErrorPageView from '$mp/components/ErrorPageView'
 import withErrorBoundary from '$shared/utils/withErrorBoundary'
-// import Analytics from '$shared/utils/Analytics'
+import Analytics from '$shared/utils/Analytics'
 import routes from '$routes'
+import ErrorBoundary from '$shared/components/ErrorBoundary'
 
 // Wrap authenticated components here instead of render() method
 // Marketplace Auth
@@ -159,40 +158,6 @@ const MiscRouter = () => ([
     <Route component={NotFoundPage} key="NotFoundPage" />,
 ])
 
-type Props = {
-    children: any,
-}
-
-type State = {}
-
-// This is a way to handle all errors in the app.
-class ErrorBoundary extends Component<Props, State> {
-    componentDidMount() {
-        if (isProduction()) {
-            console.log('prodigy')
-            Sentry.init({
-                dsn: process.env.SENTRY_DSN,
-                release: process.env.VERSION,
-                environment: process.env.SENTRY_ENVIRONMENT,
-                debug: true,
-            })
-        }
-    }
-
-    componentDidCatch(error, errorInfo) {
-        if (isProduction()) {
-            Sentry.captureException(error, {
-                extra: errorInfo,
-            })
-        }
-    }
-
-    render() {
-        const { children } = this.props
-        return children
-    }
-}
-
 const App = () => (
     <ConnectedRouter history={history}>
         <SessionProvider>
@@ -200,6 +165,7 @@ const App = () => (
                 <ModalRoot>
                     <LocaleSetter />
                     <AutoScroll />
+                    <Analytics />
                     <Switch>
                         {AuthenticationRouter()}
                         {MarketplaceRouter()}
